+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "inline-codelens" extension will be documented in thi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.2] - 2025-07-07
+
+### Added
+- Inlay Hints provider for CodeLenses.
+- Setting `inline-codelens.provider` to choose between `Decorations` and `Inlay Hints`.
+
 ## [0.0.1] - 2025-07-07
 
 ### Added